Changes to get rid of comparator triggers that weren't intended. oof. In particular, we now have an internal_disable macro that just shuts off the comparator so we don't get any interrupts latching as we're changing things

+++ src/include/libfakecatnap/events.h
@@ -26,7 +26,7 @@ typedef struct evt_ {
   evt_valid_t valid;
   uint16_t period;
   uint8_t periodic;
-  float V_safe;
+  uint16_t V_safe;
   float V_final;
   float V_min;
 } evt_t;
